test move autopatch
