+++ tesp/test_canary_integration.py
