+++ tesp/test_pre_release.py
@@ -668,139 +668,6 @@ def test_dataclasses_check(image):
         )
 
 
-@pytest.mark.usefixtures("sagemaker")
-@pytest.mark.model("N/A")
-def test_pip_check(image):
-    """
-    Ensure there are no broken requirements on the containers by running "pip check"
-
-    :param image: ECR image URI
-    """
-
-    allowed_exceptions = []
-
-    # SageMaker Python SDK updated its pyyaml requirement to 6.0, which is incompatible with the
-    # requirement from awscli. awscli only requires pyyaml for ecs/eks related invocations, while
-    # pyyaml usage seems to be more fundamental in sagemaker. Therefore, we are ignoring awscli's
-    # requirement in favor of sagemaker.
-    allowed_exceptions.append(
-        r"^awscli \d+(\.\d+)* has requirement PyYAML<5\.5,>=3\.10, but you have pyyaml 6\.0.$"
-    )
-
-    # TF inference containers do not have core tensorflow installed by design. Allowing for this pip check error
-    # to occur in order to catch other pip check issues that may be associated with TF inference
-    # smclarify binaries have s3fs->aiobotocore dependency which uses older version of botocore. temporarily
-    # allowing this to catch other issues
-    gpu_suffix = "-gpu" if "gpu" in image else ""
-    allowed_exceptions.append(
-        rf"^tensorflow-serving-api{gpu_suffix} \d\.\d+\.\d+ requires tensorflow(|{gpu_suffix}), which is not installed.$"
-    )
-
-    allowed_exceptions.append(
-        r"^aiobotocore \d+(\.\d+)* has requirement botocore<\d+(\.\d+)*,>=\d+(\.\d+)*, but you have botocore \d+(\.\d+)*\.$"
-    )
-
-    # The v0.22 version of tensorflow-io has a bug fixed in v0.23 https://github.com/tensorflow/io/releases/tag/v0.23.0
-    allowed_exceptions.append(
-        rf"^tensorflow-io 0.22.0 requires tensorflow, which is not installed.$"
-    )
-
-    framework, framework_version = get_framework_and_version_from_tag(image)
-    # The v0.21 version of tensorflow-io has a bug fixed in v0.23 https://github.com/tensorflow/io/releases/tag/v0.23.0
-
-    tf263_io21_issue_framework_list = [
-        "tensorflow",
-        "huggingface_tensorflow",
-        "huggingface_tensorflow_trcomp",
-    ]
-    if framework in tf263_io21_issue_framework_list or Version(framework_version) in SpecifierSet(
-        ">=2.6.3,<2.7"
-    ):
-        allowed_exceptions.append(
-            rf"^tensorflow-io 0.21.0 requires tensorflow, which is not installed.$"
-        )
-
-    # TF2.9 sagemaker containers introduce tf-models-official which has a known bug where in it does not respect the
-    # existing TF installation. https://github.com/tensorflow/models/issues/9267. This package in turn brings in
-    # tensorflow-text. Skip checking these two packages as this is an upstream issue.
-    if framework in ["tensorflow", "huggingface_tensorflow"] and Version(
-        framework_version
-    ) in SpecifierSet(">=2.9.1"):
-        exception_strings = []
-
-        for ex_ver in [
-            "2.9.1",
-            "2.9.2",
-            "2.10.0",
-            "2.11.0",
-            "2.12.0",
-            "2.13.0",
-            "2.14.2",
-            "2.16.0",
-        ]:
-            exception_strings += [f"tf-models-official {ex_ver}".replace(".", r"\.")]
-
-        for ex_ver in ["2.9.0", "2.10.0", "2.11.0", "2.12.0", "2.13.0", "2.14.0", "2.16.1"]:
-            exception_strings += [f"tensorflow-text {ex_ver}".replace(".", r"\.")]
-
-        for ex_ver in ["2.16.0"]:
-            exception_strings += [f"tf-keras {ex_ver}".replace(".", r"\.")]
-
-        allowed_exceptions.append(
-            rf"^({'|'.join(exception_strings)}) requires tensorflow, which is not installed."
-        )
-
-        allowed_exceptions.append(
-            rf"tf-models-official 2.9.2 has requirement tensorflow-text~=2.9.0, but you have tensorflow-text 2.10.0."
-        )
-
-    if (
-        framework in ["pytorch"]
-        and Version(framework_version) in SpecifierSet("==2.3.*")
-        and all([substr_to_find in image for substr_to_find in ["gpu", "training"]])
-    ):
-        exception_strings = []
-
-        # Adding due to the latest pip check platform feature: https://github.com/pypa/pip/issues/12884
-        for ex_ver in ["1.11.1.1"]:
-            exception_strings += [f"ninja {ex_ver}".replace(".", r"\.")]
-
-        allowed_exceptions.append(
-            rf"^({'|'.join(exception_strings)}) is not supported on this platform"
-        )
-
-    if "pytorch" in image and "trcomp" in image:
-        allowed_exceptions.extend(
-            [
-                r"torch-xla \d+(\.\d+)* requires absl-py, which is not installed.",
-                r"torch-xla \d+(\.\d+)* requires cloud-tpu-client, which is not installed.",
-            ]
-        )
-
-    if "autogluon" in image:
-        allowed_exceptions.extend(
-            [
-                r"openxlab \d+\.\d+\.\d+ has requirement requests~=\d+\.\d+\.\d+, but you have requests .*",
-                r"openxlab \d+\.\d+\.\d+ has requirement setuptools~=\d+\.\d+\.\d+, but you have setuptools .*",
-            ]
-        )
-
-    # Add null entrypoint to ensure command exits immediately
-    ctx = Context()
-    output = ctx.run(f"docker run --entrypoint='' {image} pip check", hide=True, warn=True)
-    if output.return_code != 0:
-        if not (
-            any(
-                [
-                    re.findall(allowed_exception, output.stdout)
-                    for allowed_exception in allowed_exceptions
-                ]
-            )
-        ):
-            # Rerun pip check test if this is an unexpected failure
-            ctx.run(f"docker run --entrypoint='' {image} pip check", hide=True)
-
-
 @pytest.mark.usefixtures("sagemaker", "huggingface")
 @pytest.mark.model("N/A")
 def test_cuda_paths(gpu):
@@ -942,82 +809,6 @@ def _test_sm_toolkit_and_ts_version(image, region):
     ), f"The label {expected_label} which enforces compatability between sagemaker inference toolkit and torchserve seems to be invalid/missing for the image {image}"
 
 
-@pytest.mark.usefixtures("sagemaker")
-@pytest.mark.integration("oss_compliance")
-@pytest.mark.model("N/A")
-@pytest.mark.skipif(
-    not is_dlc_cicd_context(), reason="We need to test OSS compliance only on PRs and pipelines"
-)
-def test_oss_compliance(image):
-    """
-    Run oss compliance check on a container to check if license attribution files exist.
-    And upload source of third party packages to S3 bucket.
-    """
-    THIRD_PARTY_SOURCE_CODE_BUCKET = "aws-dlinfra-licenses"
-    THIRD_PARTY_SOURCE_CODE_BUCKET_PATH = "third_party_source_code"
-    file = "THIRD_PARTY_SOURCE_CODE_URLS"
-    container_name = get_container_name("oss_compliance", image)
-    context = Context()
-    local_repo_path = get_repository_local_path()
-    start_container(container_name, image, context)
-
-    # run compliance test to make sure license attribution files exists. testOSSCompliance is copied as part of Dockerfile
-    run_cmd_on_container(container_name, context, "/usr/local/bin/testOSSCompliance /root")
-
-    try:
-        context.run(
-            f"docker cp {container_name}:/root/{file} {os.path.join(local_repo_path, file)}"
-        )
-    finally:
-        context.run(f"docker rm -f {container_name}", hide=True)
-
-    s3_resource = boto3.resource("s3")
-
-    with open(os.path.join(local_repo_path, file)) as source_code_file:
-        for line in source_code_file:
-            name, version, url = line.split(" ")
-            file_name = f"{name}_v{version}_source_code"
-            s3_object_path = f"{THIRD_PARTY_SOURCE_CODE_BUCKET_PATH}/{file_name}.tar.gz"
-            local_file_path = os.path.join(local_repo_path, file_name)
-
-            for i in range(3):
-                try:
-                    if not os.path.isdir(local_file_path):
-                        context.run(f"git clone {url.rstrip()} {local_file_path}", hide=True)
-                        context.run(f"tar -czvf {local_file_path}.tar.gz {local_file_path}")
-                except Exception as e:
-                    time.sleep(1)
-                    if i == 2:
-                        LOGGER.error(f"Unable to clone git repo. Error: {e}")
-                        raise
-                    continue
-            try:
-                if os.path.exists(f"{local_file_path}.tar.gz"):
-                    LOGGER.info(f"Uploading package to s3 bucket: {line}")
-                    s3_resource.Object(THIRD_PARTY_SOURCE_CODE_BUCKET, s3_object_path).load()
-            except botocore.exceptions.ClientError as e:
-                if e.response["Error"]["Code"] == "404":
-                    try:
-                        # using aws cli as using boto3 expects to upload folder by iterating through each file instead of entire folder.
-                        context.run(
-                            f"aws s3 cp {local_file_path}.tar.gz s3://{THIRD_PARTY_SOURCE_CODE_BUCKET}/{s3_object_path}"
-                        )
-                        object = s3_resource.Bucket(THIRD_PARTY_SOURCE_CODE_BUCKET).Object(
-                            s3_object_path
-                        )
-                        object.Acl().put(ACL="public-read")
-                    except ClientError as e:
-                        LOGGER.error(
-                            f"Unable to upload source code to bucket {THIRD_PARTY_SOURCE_CODE_BUCKET}. Error: {e}"
-                        )
-                        raise
-                else:
-                    LOGGER.error(
-                        f"Unable to check if source code is present on bucket {THIRD_PARTY_SOURCE_CODE_BUCKET}. Error: {e}"
-                    )
-                    raise
-
-
 @pytest.mark.usefixtures("sagemaker_only")
 @pytest.mark.model("N/A")
 def test_pytorch_training_sm_env_variables(pytorch_training):
@@ -1079,115 +870,3 @@ def test_mxnet_training_sm_env_variables(mxnet_training):
         env_vars_to_test=env_vars,
         container_name_prefix=container_name_prefix,
     )
-
-
-@pytest.mark.usefixtures("sagemaker")
-@pytest.mark.model("N/A")
-def test_core_package_version(image):
-    """
-    In this test, we ensure that if a core_packages.json file exists for an image, the packages installed in the image
-    satisfy the version constraints specified in the core_packages.json file.
-    """
-    core_packages_path = src_utils.get_core_packages_path(image)
-    if not os.path.exists(core_packages_path):
-        pytest.skip(f"Core packages file {core_packages_path} does not exist for {image}")
-    LOGGER.info(f"Core packages file {core_packages_path} for {image}")
-
-    with open(core_packages_path, "r") as f:
-        core_packages = json.load(f)
-
-    ctx = Context()
-    container_name = get_container_name("test_core_package_version", image)
-    start_container(container_name, image, ctx)
-    docker_exec_command = f"""docker exec --user root {container_name}"""
-    installed_package_version_dict = get_installed_python_packages_with_version(docker_exec_command)
-
-    violation_data = {}
-
-    for package_name, specs in core_packages.items():
-        package_name = package_name.lower()
-        installed_version = None
-        if package_name not in installed_package_version_dict:
-            violation_data[
-                package_name
-            ] = f"Package: {package_name} not installed in {installed_package_version_dict}"
-        else:
-            installed_version = Version(installed_package_version_dict[package_name])
-        if installed_version and installed_version not in SpecifierSet(
-            specs.get("version_specifier")
-        ):
-            violation_data[package_name] = (
-                f"Package: {package_name} version {installed_version} does not match "
-                f"requirement {specs.get('version_specifier')}"
-            )
-
-    stop_and_remove_container(container_name, ctx)
-    assert (
-        not violation_data
-    ), f"Few packages violate the core_package specifications: {violation_data}"
-
-
-@pytest.mark.model("N/A")
-def test_package_version_regression_in_image(image):
-    """
-    This test verifies if the python package versions in the already released image are not being downgraded/deleted in the
-    new released. This test would be skipped for images whose BuildSpec does not have `latest_release_tag` or `release_repository`
-    keys in the buildspec - as these keys are used to extract the released image uri. Additionally, if the image is not already
-    released, this test would be skipped.
-    """
-    dlc_path = os.getcwd().split("/test/")[0]
-    corresponding_image_spec = get_image_spec_from_buildspec(
-        image_uri=image, dlc_folder_path=dlc_path
-    )
-
-    if any(
-        [
-            expected_key not in corresponding_image_spec
-            for expected_key in ["latest_release_tag", "release_repository"]
-        ]
-    ):
-        pytest.skip(f"Image {image} does not have `latest_release_tag` in its buildspec.")
-
-    previous_released_image_uri = f"""{corresponding_image_spec["release_repository"]}:{corresponding_image_spec["latest_release_tag"]}"""
-
-    LOGGER.info(f"Image spec for {image}: {json.dumps(corresponding_image_spec)}")
-    ctx = Context()
-
-    # Pull previous_released_image_uri
-    prod_account_id = get_account_id_from_image_uri(image_uri=previous_released_image_uri)
-    prod_image_region = get_region_from_image_uri(image_uri=previous_released_image_uri)
-    login_to_ecr_registry(context=ctx, account_id=prod_account_id, region=prod_image_region)
-    previous_released_image_uri = f"{previous_released_image_uri}"
-    run_output = ctx.run(f"docker pull {previous_released_image_uri}", warn=True, hide=True)
-    if not run_output.ok:
-        if "requested image not found" in run_output.stderr.lower():
-            pytest.skip(
-                f"Previous image: {previous_released_image_uri} for Image: {image} has not been released yet"
-            )
-        raise DockerImagePullException(
-            f"{run_output.stderr} when pulling {previous_released_image_uri}"
-        )
-
-    # Get the installed python package versions and find any regressions
-    current_image_package_version_dict = get_installed_python_packages_using_image_uri(
-        context=ctx, image_uri=image
-    )
-    released_image_package_version_dict = get_installed_python_packages_using_image_uri(
-        context=ctx, image_uri=previous_released_image_uri
-    )
-    violating_packages = {}
-    for package_name, version_in_released_image in released_image_package_version_dict.items():
-        if package_name not in current_image_package_version_dict:
-            violating_packages[
-                package_name
-            ] = "Not present in the image that is being currently built."
-            continue
-        version_in_current_image = current_image_package_version_dict[package_name]
-        if Version(version_in_released_image) > Version(version_in_current_image):
-            violating_packages[
-                package_name
-            ] = f"Version in already released image: {version_in_released_image} is greater that version in current image: {version_in_current_image}"
-
-    assert (
-        not violating_packages
-    ), f"Package regression observed between already released image: {previous_released_image_uri} and current image: {image}. Violating packages: {json.dumps(violating_packages)}"
