+++ test/dlc_tests/sanity/test_binary_visibility.py
