+++ test/dlc_tests/sanity/test_scan_vulnerability_class_functionality.py
