+++ test/dlc_tests/sanity/test_telemetry.py
